estimators: delete loss_config.h, fold LossConfig into global_positioning.h
Follow-up to D1 dedup. The loss_config.h after the rewrite was just a
~25-line struct definition with one consumer (GlobalPositionerOptions).
Inline it into global_positioning.h next to PointConstraintType so the
GP-specific helpers live together.

- LossConfig struct + CreateLossFunction() method moved to
  colmap/estimators/global_positioning.h.
- loss_config.h deleted; CMakeLists.txt entry removed.
- Includers (global_positioning.cc, motion_averaging.cc binding) drop
  the include — global_positioning.h covers it transitively.

Net: 1 fewer header in the tree.
